refactor(functions): firestoreへのアクセスを別ファイルで管理 #72
