+++ functions/src/index.ts
@@ -2,3 +2,5 @@ export { createUser } from './user.function';
 export { incrementStoreViewCount } from './store.function';
 export { deleteUser } from './user.function';
 export { deleteUserData } from './user.function';
+export { createStripeCustomer } from './stripe/customer.function';
+export { getStripeCustomer } from './stripe/customer.function';
